All of 0.3.6.+, currently in alphas.

Going to be attempting 0.3.6 for 1.18 firstly before working more into 0.3.+
This push is to help reset the work stack for the team as they continue to add content to 1.16, until 1.18 is working well with full feature parody with the latest 0.3.6.+ version.

+++ src/main/java/com/tac/guns/crafting/WorkbenchRecipeSerializer.java
@@ -59,7 +59,7 @@ public class WorkbenchRecipeSerializer extends net.minecraftforge.registries.For
         int size = buffer.readVarInt();
         for(int i = 0; i < size; i++) {
             Ingredient ingredient = Ingredient.read(buffer);
-            int count = buffer.readByte();
+            int count = buffer.readInt();
             builder.add(new Pair<>(ingredient, count));
         }
         return new WorkbenchRecipe(recipeId, result, builder.build(), group);
@@ -74,7 +74,7 @@ public class WorkbenchRecipeSerializer extends net.minecraftforge.registries.For
         for(Pair<Ingredient, Integer> stack : recipe.getMaterials())
         {
             stack.getFirst().write(buffer);
-            buffer.writeByte(stack.getSecond());
+            buffer.writeInt(stack.getSecond());
         }
     }
 }
